refactor(work-management): lift the room-definition contract and its resolvers out of the registry (BI-EFFD97B4)
source-registry.ts crossed the 800-line module ceiling once every definition carried a trigger, grant ceiling and measure. The vocabulary (trigger kinds, tool grant, measure, narrowRoomToolGrant) now lives in room-definition-contract.ts, the two registry-backed resolvers in room-definition-resolvers.ts; the registry re-exports the contract so existing imports keep working.
